Fixed a bug that discarded a struct from send buffer in control everytime something was added if there already were anything in the buffer. Some small changes and bug fixes in other places as well. Position seems to become better and better, but is still far from perfect.

+++ RobotStyr2.0/RobotStyr2.0/map.c
@@ -1,10 +1,10 @@
 #include "map.h"
 
 // MAP variables
-int x = 0;
-int y = 0;
-int xdir = 0;
-int ydir = 1;
+int8_t x = 0;
+int8_t y = 0;
+int8_t xdir = 0;
+int8_t ydir = 1;
 
 
 //Shifts the elements up. (Keep in mind that lower left corner is the origin)
